+++ installation.sh
@@ -59,9 +59,9 @@ if [ -z "$VERSION" ]; then
 fi
 
 RELEASE_URL="https://github.com/$REPO/releases/download/$VERSION/gysmo"
-CONFIG_URL="https://raw.githubusercontent.com/$REPO/refs/tags/$VERSION/src/config/config.json"
-SCHEMA_URL="https://raw.githubusercontent.com/$REPO/refs/tags/$VERSION/src/config/schema/config_schema.json"
-ASCII_URL="https://raw.githubusercontent.com/$REPO/refs/tags/$VERSION/src/ascii/gysmo"
+CONFIG_URL="https://raw.githubusercontent.com/$REPO/refs/tags/$VERSION/gysmo/config/config.json"
+SCHEMA_URL="https://raw.githubusercontent.com/$REPO/refs/tags/$VERSION/gysmo/config/schema/config_schema.json"
+ASCII_URL="https://raw.githubusercontent.com/$REPO/refs/tags/$VERSION/gysmo/ascii/gysmo"
 
 # Create the necessary directory structure in ~/.config/gysmo
 echo "Setting up default templates in $HOME/.config/gysmo..."