+++ flake.nix
@@ -14,7 +14,7 @@
       {
         packages.default = pkgs.buildGoModule {
           pname = "gysmo";
-          version = "0.2.1";
+          version = "0.2.2";
           src = self;
           vendorHash = null;
         };